Consolidate launcher HTTP helpers
